Fixed tense consistency (used past) in statuses
In several places in Watcher, there are statuses which corresponds
to a status in the respective life cycle. There were used past tense
everywhere, except "SUCCESS" which was written in present tense.

So I searched for all the occurences of this state, and replaced them
with their past tense form.

Change-Id: Ic41def96d1a608da06a1e394e24704d4f06ab32a
Closes-Bug: #1520276