Removed py26 support
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support from
their libraries. sine we rely on oslo for a lot of our work, and depend
on infra for our CI, we should drop py26 support too.

Change-Id: I8aec07a1149ad7232c3d6db2f3c274ec9ce11a66
Closes-Bug: #1522363