Remove bandit from lower-constraints
The version of bandit in lower-constraints (1.4.0) does
not match the version in test-requirements (1.6.0) however
bandit is a test-only dependency and there is no test coverage
for bandit in the lower-constraints tox job target, so there
is really no good reason to have bandit in lower-constraints.
As such, this change simply removes it from lower-constraints.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Change-Id: I35f66994e9a3a334b342232587d84491542da755