Configure nova notification_format for grenade
Nova changed the default notification_format from "both" to
"unversioned" in Train [1]. Without configuring nova in the
grenade job we are not testing the nova versioned notification
handler code during upgrades.

Note that grenade only runs stack.sh on the base (old) side so
this change has to depend on a devstack stable/stein change to
add the NOVA_NOTIFICATION_FORMAT variable that we override.

Closes-Bug: #1831917

Depends-On: Ied9d50b07c368d5c2be658c744f340a8d1ee41e0

[1] https://review.opendev.org/603079/

Change-Id: I94c2d14477da185310e0fec596a1ad6436b802f1